Make inspection a bounded context with a pure core
composition/inspect.ts held the file-search toolkit in one 508-line
file: file selection, text matching, JSON path parsing, JSON document
queries, locators, occurrence selection, and the filesystem calls that
feed them. Every rule was only reachable through files on disk.

Move it to inspect/, the language shared by Checks and mutations.
inspect/core holds selection parts and exclusion, text matching with
locations, the JSON path parser and document query, locators and
occurrence selection, and the search Scan with timestamps passed in.
inspect/shell holds the file operations and the searches that read
files and the clock. The text locator now shares one occurrence rule
with the JSON locator. Public names and messages are unchanged.
