+++ gates/support/effects-model.ts
@@ -30,7 +30,8 @@ const APPROVED_EFFECT_BOUNDARIES = new Set([
   'packages/redproof/src/cli/shell/main.ts',
   'packages/redproof/src/command/shell/check.ts',
   'packages/redproof/src/command/shell/spawn.ts',
-  'packages/redproof/src/composition/inspect.ts',
+  'packages/redproof/src/inspect/shell/files.ts',
+  'packages/redproof/src/inspect/shell/search.ts',
   'packages/redproof/src/composition/mutation/context.ts',
   'packages/redproof/src/composition/mutation/filesystem.ts',
   'packages/redproof/src/reporter/shell/sources.ts',
