+++ packages/redproof/src/inspect/index.ts
@@ -1,0 +1,11 @@
+export type { FileSelection, FindFilesOptions } from './core/selection.ts';
+export type { ScanOptions, SearchScan } from './core/scan.ts';
+export type { FindTextOptions, TextMatch } from './core/text.ts';
+export type { FindJsonOptions, JsonMatch, JsonPathToken } from './core/json.ts';
+export { parseJsonPath } from './core/json.ts';
+export type { JsonLocator, JsonOccurrence, TextLocator, TextOccurrence } from './core/locate.ts';
+export { locate } from './core/locate.ts';
+export type { RootContext } from './shell/files.ts';
+export { files } from './shell/files.ts';
+export type { JsonSearch, TextSearch } from './shell/search.ts';
+export { json, resolveJsonLocator, resolveTextLocator, text } from './shell/search.ts';
